% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel_order.R
\name{cancel_order}
\alias{cancel_order}
\title{Cancel Specific Order by ID for Account}
\usage{
cancel_order(tokens, encrypted_account_id, order_id)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}

\item{encrypted_account_id}{encrypted ID of the account from `get_account_numbers` function (string).}

\item{order_id}{order ID for the order (numeric).}
}
\value{
Returns a message informing the user if cancellation
        was successful or if there was an error.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function, the encrypted account ID, and the order ID, cancel
the specific order.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, July 2024
}
\keyword{ID}
\keyword{account}
\keyword{cancel}
\keyword{order}
