% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_numbers.R
\name{get_account_numbers}
\alias{get_account_numbers}
\title{Get Account Numbers}
\usage{
get_account_numbers(tokens)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}
}
\value{
Returns a data frame containing the account numbers
        and their encrypted values.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function, return the account number(s) information, including
the actual account number(s) and the encrypted ID(s) of the user
that were granted when authenticating. The encrypted IDs are used
in other functions (like orders and transactions) that require a
specific encrypted account ID to be specified.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, June 2024
}
\keyword{account}
\keyword{encrypted}
\keyword{numbers}
