% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accounts.R
\name{get_accounts}
\alias{get_accounts}
\title{Get Accounts Information}
\usage{
get_accounts(tokens, fields = NULL)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}

\item{fields}{specific fields to be returned, an example being "positions" (string).}
}
\value{
Returns a data frame containing the account(s)
        information.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function, return the account(s) information associated with
the authenticated user. By default, this includes positions,
fundamentals, and general account information. However, one
can use the `fields` argument to get more specific as to the
information returned.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, June 2024
}
\keyword{accounts}
\keyword{positions}
