% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instruments.R
\name{get_instruments}
\alias{get_instruments}
\title{Get Instruments}
\usage{
get_instruments(tokens, symbol, projection)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}

\item{symbol}{symbol(s) of interest to search for (string or character vector).}

\item{projection}{type of search to be performed. Valid values are "symbol-search", "symbol-regex", "desc-search", "desc-regex", "search", and "fundamental" (string).}
}
\value{
Returns a data frame containing information surrounding
        the symbol(s) of interest in the search.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function, the symbol(s) of interest, and the search type, return
a data frame with information about the securities matching
the search.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, July 2024
}
\keyword{instrument}
\keyword{instruments}
\keyword{search}
