% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instruments_cusip.R
\name{get_instruments_cusip}
\alias{get_instruments_cusip}
\title{Get Instruments by Cusip}
\usage{
get_instruments_cusip(tokens, cusip_id)
}
\arguments{
\item{tokens}{tokens object from `get_authentication_tokens` function (list).}

\item{cusip_id}{cusip of the security to be searched on (string).}
}
\value{
Returns a data frame containing information surrounding
        the cusip of interest in the search.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function and a cusip, return a data frame with information
about the security matching the search.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, July 2024
}
\keyword{cusip}
\keyword{instrument}
\keyword{instruments}
\keyword{search}
