% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_hours.R
\name{get_market_hours}
\alias{get_market_hours}
\title{Get Market Hours}
\usage{
get_market_hours(
  tokens,
  markets = c("equity", "option", "bond", "future", "forex"),
  date = NULL
)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}

\item{markets}{markets of interest that are "equity", "option", "bond", "future", and/or "forex". Default is all markets (string or character vector).}

\item{date}{date you would like to get the hours of operation. Valid dates are today through one year from now. Default is NULL, which is today (date).}
}
\value{
Returns a data frame containing information surrounding
        the market of interest and its specific hours of operation.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function, return a data frame containing information about the
market(s) of interest and its specific hours of operation. By
default, all of the markets are returned for today's date, but
both the specific markets returned and the date can be tweaked.
Please see the parameters for more specifics related to what can
be specified for the function.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, July 2024
}
\keyword{hours}
\keyword{market}
