% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_hours_single.R
\name{get_market_hours_single}
\alias{get_market_hours_single}
\title{Get Market Hours for Single Market}
\usage{
get_market_hours_single(tokens, market, date = NULL)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}

\item{market}{market of interest that is either "equity", "option", "bond", "future", or "forex" (string).}

\item{date}{date you would like to get the hours of operation. Valid dates are today through one year from now. Default is NULL, which is today (date).}
}
\value{
Returns a data frame containing information surrounding
        the market's hours of operation.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function and the market of interest, return specific information
about that market's hours of operation. Note that a data parameter
can also be specified between now and one year from now to obtain
specific hours of operation for the market related to that date.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, July 2024
}
\keyword{hours}
\keyword{market}
\keyword{operation}
\keyword{single}
