% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_option_expiration_chain.R
\name{get_option_expiration_chain}
\alias{get_option_expiration_chain}
\title{Get Option Expiration Chain}
\usage{
get_option_expiration_chain(tokens, symbol)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}

\item{symbol}{symbol for option expiration chain (string).}
}
\value{
Returns a data frame containing information surrounding
        the option expiration chain.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function and the symbol of interest, return the option expiration
chain information related to the symbol. This includes expiration
dates, expiration types, settlement types, and more.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, June 2024
}
\keyword{chain}
\keyword{expiration}
\keyword{option}
