% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_order_id.R
\name{get_order_id}
\alias{get_order_id}
\title{Get Specific Order by ID for Account}
\usage{
get_order_id(tokens, encrypted_account_id, order_id)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}

\item{encrypted_account_id}{encrypted ID of the account from `get_account_numbers` function (string).}

\item{order_id}{order ID for the order (numeric).}
}
\value{
Returns a data frame containing the information about
        the specific order.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function, the encrypted account ID, and the order ID, return
a data frame containing information about the specific order.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, July 2024
}
\keyword{ID}
\keyword{account}
\keyword{order}
