% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quotes.R
\name{get_quotes}
\alias{get_quotes}
\title{Get Quotes}
\usage{
get_quotes(tokens, symbols, fields = NULL, indicative = NULL)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}

\item{symbols}{symbols for quotes (vector).}

\item{fields}{request for subset of data. Possible values are NULL, "all", "quote", "fundamental", "extended", "reference", or "regular". Note these can be combined in a vector or a string with a value can be used (string or vector).}

\item{indicative}{include indicative symbol quotes or not for ETF requests. Possible values are "true" and "false" (string).}
}
\value{
Returns a data frame containing information about the
        given symbols of interest.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function and the symbols of interest, return a data frame
containing information about those symbols. Note that this
function can return information that goes beyond a standard
quote (for example, fundamental information can be returned).
#' By default, everything is returned, but the specific information
returned can be customized through the `fields` argument below.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, June 2024
}
\keyword{quote}
\keyword{quotes}
\keyword{symbols}
