% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quotes_single_symbol.R
\name{get_quotes_single_symbol}
\alias{get_quotes_single_symbol}
\title{Get Quotes for Single Symbol}
\usage{
get_quotes_single_symbol(tokens, symbol_id, fields = NULL)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}

\item{symbol_id}{symbol for the security of interest}

\item{fields}{request for subset of data. Possible values are NULL, "all", "quote", "fundamental", "extended", "reference", or "regular". Note these can be combined in a vector or a string with a value can be used. The default is NULL, which is everything (string or vector).}
}
\value{
Returns a data frame containing information about the
        given symbol of interest.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function and the symbol ID, return a data frame
containing information about that symbol. Note that this
function can return information that goes beyond a standard
quote (for example, fundamental information can be returned).
By default, everything is returned, but the specific information
returned can be customized through the `fields` argument below.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, June 2024
}
\keyword{quote}
\keyword{symbol_id}
