% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_preferences.R
\name{get_user_preferences}
\alias{get_user_preferences}
\title{Get User Preferences}
\usage{
get_user_preferences(tokens)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}
}
\value{
Returns a data frame containing the user's preferences
}
\description{
Given the tokens object from the `get_authentication_tokens`
function, returns the user preferences associated with the
account that was authenticated.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, June 2024
}
\keyword{accounts}
\keyword{preferences}
