% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account.R
\name{get_account}
\alias{get_account}
\title{Get Account Information for Specific Account}
\usage{
get_account(tokens, encrypted_account_id, fields = NULL)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}

\item{encrypted_account_id}{encrypted ID of the account from `get_account_numbers` function (string).}

\item{fields}{specific fields to be returned, one example being "positions" (string).}
}
\value{
Returns a data frame containing the account
        information. This includes position information
        a day trader flag, the account number, and more.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function and the encrypted account ID, return the account
information (positions, fundamentals, and general account
information). The encrypted account ID can be found using the
`get_account_numbers` function.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, July 2024
}
\keyword{account}
\keyword{number}
\keyword{positions}
