% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transaction.R
\name{get_transaction}
\alias{get_transaction}
\title{Get Specific Transaction Information}
\usage{
get_transaction(tokens, account_number, transaction_id)
}
\arguments{
\item{tokens}{token object from `get_authentication_tokens` function (list).}

\item{account_number}{encrypted ID of the account (string).}

\item{transaction_id}{transaction ID of interest (numeric).}
}
\value{
Returns a data frame containing the transaction
        information.
}
\description{
Given the tokens object from the `get_authentication_tokens`
function, the encrypted account ID of interest, and the
transaction ID of interest, return information corresponding
to that transaction. Note that the transaction ID can be obtained
by first calling the `get_transactions` function, finding the transaction
of interest, and then finding the transaction ID on that data frame.
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, July 2024
}
\keyword{account}
\keyword{transaction}
